requirements.txt
pillow~=9.5.0
opencv-python~=4.7.0.72
numpy~=1.24.3
clip-openai
requests
selenium~=4.9.0
cryptocode~=0.1
imagehash~=4.3.1
nicegui~=1.2.11
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.13.1+cu116